Collapse things more
remove provider concept, inline the webauthn code to oidc server
