+++ storage.go
@@ -6,6 +6,25 @@ import (
 	"github.com/pardot/oidc/core"
 )
 
+// Authentication are the details flagged for an authenticated user of the
+// system.
+type Authentication struct {
+	// Subject (required) is the unique identifier for the authenticated user.
+	// This should be stable over time.
+	Subject string `dynamodbav:"subject"`
+	// EMail (optional), for when the email/profile scope is requested
+	EMail string `dynamodbav:"email,omitempty"`
+	// FullName (optional), for when the profile scope is requested
+	FullName string `dynamodbav:"full_name,omitempty"`
+	// Groups (optional), for when the groups scope is requested
+	Groups []string `dynamodbav:"groups,omitempty"`
+	// ExtraClaims (optional) fields to add to the returned ID token claims
+	ExtraClaims map[string]interface{} `dynamodbav:"extra_claims,omitempty"`
+	// PolicyContext is internal data, that is passed to the policies that are
+	// evaluated downstream. This data is not presented to the user.
+	PolicyContext map[string]interface{} `dynamodbav:"policy_context,omitempty"`
+}
+
 type Storage interface {
 	core.SessionManager
 	WebauthnUserStore
