+++ web/templates/login.tmpl.html
@@ -1,0 +1,121 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <title>title</title>
+    <link rel="stylesheet" href="style.css">
+    <script src="script.js"></script>
+</head>
+
+<body>
+
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
+    <script>
+
+        class WebAuthn {
+            // Decode a base64 string into a Uint8Array.
+            static _decodeBuffer(value) {
+                return Uint8Array.from(atob(value), c => c.charCodeAt(0));
+            }
+
+            // Encode an ArrayBuffer into a urlbase64 string.
+            static _encodeBuffer(value) {
+                return btoa(String.fromCharCode.apply(null, new Uint8Array(value)))
+                    .replace(/\+/g, "-")
+                    .replace(/\//g, "_")
+                    .replace(/=/g, "");;
+            }
+
+            // Checks whether the status returned matches the status given.
+            static _checkStatus(status) {
+                return res => {
+                    if (res.status === status) {
+                        return res;
+                    }
+                    throw new Error(res.statusText);
+                };
+            }
+
+            login() {
+                let email = $("#email").val()
+                if (email === "") {
+                    alert("Please enter a email");
+                    return;
+                }
+
+                return fetch('/start?email=' + email + '&sessionID={{ SessionID }}', {
+                    method: 'POST',
+                    body: JSON.stringify({}) // TODO - better than query string
+                })
+                    .then(WebAuthn._checkStatus(200))
+                    .then(res => res.json())
+                    .then(res => {
+                        res.publicKey.challenge = WebAuthn._decodeBuffer(res.publicKey.challenge);
+                        if (res.publicKey.allowCredentials) {
+                            for (let i = 0; i < res.publicKey.allowCredentials.length; i++) {
+                                res.publicKey.allowCredentials[i].id = WebAuthn._decodeBuffer(res.publicKey.allowCredentials[i].id);
+                            }
+                        }
+                        return res;
+                    })
+                    .then(res => navigator.credentials.get(res))
+                    .then(credential => {
+                        return fetch('/finish?email=' + email + '&sessionID={{ SessionID }}', {
+                            method: 'POST',
+                            headers: {
+                                'Accept': 'application/json',
+                                'Content-Type': 'application/json'
+                            },
+                            body: JSON.stringify({
+                                id: credential.id,
+                                rawId: WebAuthn._encodeBuffer(credential.rawId),
+                                type: credential.type,
+                                response: {
+                                    clientDataJSON: WebAuthn._encodeBuffer(credential.response.clientDataJSON),
+                                    authenticatorData: WebAuthn._encodeBuffer(credential.response.authenticatorData),
+                                    signature: WebAuthn._encodeBuffer(credential.response.signature),
+                                    userHandle: WebAuthn._encodeBuffer(credential.response.userHandle)
+                                }
+                            }),
+                        })
+                    })
+                    .then(WebAuthn._checkStatus(200));
+            }
+        }
+
+        let w = new WebAuthn();
+
+        let loginPending = false;
+
+        function doLogin() {
+            if (loginPending) return;
+            loginPending = true;
+            // document.getElementById("loginLoading").classList.remove("hide");
+            w.login({ "username": document.getElementById("email").value })
+                // .then(res => res.json())
+                .then(res => {
+                    window.location.href = '/loggedin';
+                })
+                .catch(err => {
+                    console.error(err);
+                    alert('Failed to login: ' + err);
+                })
+                .then(() => {
+                    loginPending = false;
+                    // document.getElementById("loginLoading").classList.add("hide");
+                });
+        }
+
+    </script>
+
+    <div>
+        <h3>webauthn</h3>
+        <label for="email">Email:</label>
+        <input id="email" name="email" type="text">
+        <button onclick="doLogin()">Login</button>
+    </div>
+
+</body>
+
+</html>
